[#1015] Move EventBusAuthenticationService class to top level.
In preparation for adding a unit test for that class.
